requirements.txt
beautifulsoup4==4.14.3
customtkinter==5.2.2
lxml==6.0.2
matplotlib==3.10.8
Pillow==12.1.1
python-dotenv==1.2.2
Requests==2.32.5
